mappola branch with verse changes
Started from apparatus branch, merged selectively changes from the latest main and then proceeded to making the changes required for Mappola and test them on the provided examples. This includes also a temptative classDecl file with a metDecl, which is however not valid as it uses metSym into metDecl. a listPrefixDef has been added there as well. this branch is not the one currently used by the project, and has not been tested yet in the application

Co-Authored-By: ChiaraC88 <61389529+ChiaraC88@users.noreply.github.com>